Fix hogp mouse reconnect fail
[Description]
When the hogp mouse is connecting, upper will initiate
an sdp search. This process creates a br/edr connection,
but it is rejected by the hogp mouse.
In case of a failed connection, the stack will add the
br/edr type to the device type. This causes the stack to
consider the remote device as a dual mode device.

[Root Cause]
Stack consider the remote device as a dual mode device.

[Solution]
only set the mode when creating connection succesfully

Bug: 313772305
Test: atest net_test_btif pass

Change-Id: I191d7b24e861f2246aa2150eb9010d63f917f3d8