Floss: wait when audio profile connection received but SDP is empty
In some rare cases, we could be receiving audio profile connections
while the headset has not reported its available set of audio profiles.
This may cause some profiles to be missed in propagating to the audio
server because an empty set of available profiles implies all profiles
have been connected.

In this CL, building on the observation that in such cases the headset
will eventually initiate all desired connections, we will wait for 6s of
timeout (or if both A2DP and HFP are connected) before propagating the
device to the audio server.

Bug: 289015291
Tag: #floss
Test: atest bluetooth_test_gd
Test: manually pair/forget/connect/disconnect airpods/XM3/4/5
Change-Id: If4bc92f9525099ce9c5bfc85c5a5b4fdf4905c11