[LE Audio] Handle later join case on offload
bluetooth.leaudio.dual_bidirection_swb.supported set to false is to
disable on both software and offload path in the system perspective.

When it turns to true, it means system supports this feature (and
software is default to be supported based on current design).

However offload path might not support due to limitation. Under this
condition, going through the offload path, for dual device it is not
swb but when one device is disconnected, it is swb, which makes user have
bad user experience. This patch fixed this.

Bug: 301358026
Test: manual
Test: atest -c --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_le_audio_codec_manager_test  --no-bazel-mode
Change-Id: I4b46d504963e69190d02ce0e4201ecad4fce35a0