Fix service discovery on dual mode device, when pairing on LE transport
Currently, if we find services associated with pseudo address, we would
not check if there are any services associated with static address.
This causes us to skip Classic services - they are not returned to Java,
and profiles don't get connected.

Test: manual. This issue occurs when dual mode device is bonded on LE
transport first, if it used RPA address for initial connection.
Bug: 300569218

Change-Id: Iaa08a56a4ca958820a0e5bea841866dc7148856a