Revert "LeAudioService: Do not block API for setVolume call"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/df61da59-1271-4b00-b5e0-38a01856d7f9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia99c733869c6b45e3cc1d0b4bbcc527f3a69b5cb