Alternating packets queued in the audio buffer
Problem: BTIF a2dp sink receives the media data and enqueues the
audio packet.It's possible that the audio buffer can become full
before being flushed. Once, the buffer is full the enqueue_buffer
the stack starts dropping the old packets. However, it doesn't
queue the incoming packet. This results in the packet to be
dropped alternatively and degrading the audio quality.

Solution: First queue up the incoming packet. Once the queue is
full, drop the last packet to make space for the next packet.
This will ensure there is always a space in the next incoming
packet and alternate packets are not dropped.

Bug: 279956598
Test: Tested on the seahawk device by queueing the packet.
Tag: #stability
Change-Id: Ie2377ee47de56ee2fa601f85ab363c02ea02e61c