Remove entire storage section when the device is unbonded
Every bonded device property is stored under a section corresponding to
the device. BTIF storage function for removing the device removes the
individual properties. However as the properties are added in the BT
stack, chances of some properties left alone during device removal
increases.
For example, "AddrType" property is not removed when a device is
removed. As a result, when reconnecting/repairing with such device, this
old property unintentionally gets assigned to the device.

Test: m com.android.btservices
Bug: 298479810
Change-Id: I5c1bceb708727e17771cb222abe79f986e55a573