leaudio: Update metadata after stream reconfiguration
When reconfiguring stream, state machine goes to CONFIGURED state with
certain configuration.
After that it is AF resuming stream and if stream configuration matches,
CIG is created and state machine moves to QoS Configured state.
With this path, the new applied context types and ccid is not used which
is an error.

This patch fixes that
Bug: 302481144
Test: atest bluetooth_le_audio_client_test bluetooth_le_audio_test
Tag: #feature

Change-Id: I6e1b8f09f5143a80243bc29624012526b23018f4