Unify volume conversion method and add force vol update
The volume conversion between AVRCP (127) and system volume
was originally done with floor/ceil but a change made it so
it was both floor/ceil and round in some places making the
conversion not consistent.
This change makes the conversion use round everywhere which
makes more sense as this will result in the less sound jump
as possible.

Adding a force_send parameter to device#SetVolume, in order
to be able to bypass the "volume is the same" check on
device connection so the volume is in sync with remote.

Bug: 302322731
Tag: #feature
Test: atest net_test_avrcp
Change-Id: I56cd10b04669a9daba389d612b978402dcb0a80e