Revert "leaudio: Fix for invalid offloader behavior on localSink"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8567497-5112-45ef-86f3-f1080df223e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3de1272ec11f42eb92156c17f05389cc23853d2