Unify volume conversion method and add force vol update
Upon connection, the remote device sends its volume to the
phone, and the phone will then send its own if it's not the
same. In come cases, the remote device doesn't send it's
volume and we end up not sending the local either because
it's the same as before disconnection.

This change resets the volume on disconnection and replaces
the floor/ceil used for conversion from system to AVRCP vol
by rounds, to have a better approximation.

Bug: 302322731
Tag: #feature
Test: atest net_test_avrcp
Change-Id: I56cd10b04669a9daba389d612b978402dcb0a80e