Revert "RootCanal: Harden the implementation for Set Event Filter"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/193dee3b-bc57-4370-ba9a-6b49ecd8e5b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2e20254c6f2f142ba2f07d162cfd5f2613f8056