Revert "RootCanal: Harden the implementation for Set Event Filter"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ee9015d-d478-4589-8706-f6180f4cb7db). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I01028413a40af941ed2ab0ce8687876d6e39ae3f