Revert "RootCanal: Harden the implementation for Set Event Filter"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6194eef-ad96-4990-afb5-d06a9327177b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e6476643c827c20bb14133b5503e0807fe6bd2b