LE Audio Allowlist refactor for default connection
The patch changes the LE Audio Allowlist usage to allow the default
connection perference. If the device is in the allow list, it would use
LE audio as default conneciton after paring; Otherwise, the classic
profile would be connected. Bypass LE Audio Allowlist is to allow use LE
audio connection by default for all the devices.

Bug: 300012501
Test: atest BluetoothInstrumentationTests
Change-Id: I4566537629bb2b1815ae3924a691087ea39d161b