Revert "Floss: Add address type to device property"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/caf1102a-28c0-4270-b446-a3b1e4f405c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27c0460c1700a84bfaf6b97753c77cb7d73bc496