Revert "IA2: Split btu event and btu task"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d270f936-b8ce-41b4-888a-8fdd31153373). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d4a42880d1fc1612be2d28e77c996dc986f2de5