floss: hcidoc: Refactor OddDisconnectionRule
Most rules detect the type of command or event in the process()
function, then pass them (or their parameters) to a smaller, more
specific functions. However, OddDisconnectionRule mixes these
responsibilities so it's hard to follow what one part is doing.

Refactor the rule to follow the general style.

Bug: 293397988
Tag: #floss
Test: m -j
Change-Id: I41958f63e602b9b788835fd89b1221f7515b3d4e