floss: hcidoc: fix the disconnection due to host power off
If BT is powered off, it is possible that the controller doesn't
wait until the peer device reply the disconnection request.
Therefore in aosp/2651792 we treat those disconnection request as
a special case. However, sometimes the event will still be replied,
thus we process the disconnection twice.

This CL fixes that by first checking whether the handle was
disconnected due to host power off.

Bug: 293397988
Tag: #floss
Test: Verify report 91229383170 doesn't report disconnection for
      nonexistent connection.
Test: m -j
Change-Id: I1db61463d1d93dd0120951c3a7b514d0167befd6