Prefer connected transport for service discovery on dual mode devices
If an application requests service discovery on a dual mode device with a public address without explicit preference for the transport, BT stack always prefers BR/EDR transport.
If the device is already connected over the LE transport, but not on the BR/EDR transport, then service discovery should be performed over LE transport.

Test: m com.android.btservices
Bug: 280456420

Change-Id: I43cbec21798be0fea2e23e6ecc550b75e9561f8f