Revert "floss: Restore HCI enabled state on new HCI presence"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82cf87ce-4561-4846-90c6-f3b7e349479c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If31c4146887ed67ef673d65f07a18a4819e6ca13