Check AIDL transport in LE audio HAL verifier
Unicast AIDL interface is supported after the HIDL 2.1. So it implies to
be supported for all AIDL version. Check the AIDL transport first that
could decouple the relation with actual version. It could avoid the crash as
the version upgrade which is missing to add version into bluetooth stack.
No all HAL upgrade is relates to LE audio. It's risky to require the
version need to be added each time.

Bug: 307397549
Test: make build
Change-Id: I435938cfa31ff3cb2bd211b6db151cc5123b262b