LeAudio: Fix flaky unit tests
Since the audio context notifications on the other thread invalidate the
existing audio set configurations. We should wait for the main loop to
finish the job before we start to use these configurations.

We are not consistent in every test with the synchronization after the
injected notification. This patch moves the synchronization to a
notification sending helper function, eliminating the need for new tests
to call the synchronization routine manually.

Bug: 302775890
Test: atest --host UnicastTest#StartStream_AvailableContextTypeNotifiedLater --rerun-until-failure=1000
Change-Id: I38a892c829779616ac20b0dbb283dcef5204c4e0