Migrate bluetooth-test-util-lib from cts to p/m/B
Currently, the framework SDK level Bluetooth test lib lives in CTS.
And every user of this library has to depend on CTS project. However,
not every user of this library is CTS related.

Hence, we want to move it to a shared location in p/m/B

Note: PREUPLOAD check is omited because it is a pure migration

Bug: 307752636
Test: m bluetooth-test-util-lib
Change-Id: I7d7b7f1d6a64cd11012650036ae4227e015c0dcc