Add support for UUID based IOPs
Add support for matching on UUID for LE devices. Also update config
parsing logic to use strings intead of C constructs as it's simpler and
less error prone to maintain.

Support for multiple constraints (e.g. match on name AND UUID) will be
added in a follow up CL at which point the fake IOP field added in this
CL for testing can be removed.

Bug: 290392842
Test: atest net_test_device
Change-Id: I6c067256c888d1d75501bbe8858799f57b032bef