When encryption fails due to key missing, try pairing again.
Currently, if the remote device responds with key missing during encryption, the BLE link is maintained and data can be transmitted without encryption. The attempt to encrypt in the master means that encryption was intentionally performed. However, it does not notify to the upper layer that this encryption has failed.

Therefore, the app may transmit important data without being encrypted. This is a dangerous security situation.

Test:
1. Pairing two devices
2. Remove central device on peripheral
3. Connect peripheral using cental

Change-Id: If2b53182ba59ad6a94f593fed60960bd999f80ed