floss: hcidoc: Introduce the hcidoc_packets crate
The current pdl-compiler doesn't work well with l2cap_packets.pdl.
Therefore, we temporarily write our own l2cap_packets.pdl, which is a
stripped down version of the original pdl file, but works with the
pdl-compiler. This is put into it's own crate. If we found another
non-compatible pdl files, we can utilize this crate again.

Ideally this should be obsoleted once pdl-compiler is fixed.

Bug: 306108678
Test: m -j
Test: ./build.py --target=utils, verify l2cap_packets.rs is generated.

Change-Id: I56acff17e903d6ca435c638257c8784619d5bc47