Revert "Health based action: Enable first native flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da6a5b07-58b2-4d0a-a14f-3bc2601a7148). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4279cd56d14ec61efc47d5c3c5d9e83490d0b1d2