floss: btmanagerd: Fix the state machine for Restart
To make sure the new config is applied, we should also restart
btadapterd for TurningOn state.

This patch also fixed an issue that when logging the result inside
mainloop we always assert the next states for StopBluetooth/
BluetoothStopped wrongly. It's now fixed by taking the return value
of the action functions.

Bug: 270447662
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: SetLLPrivacy on On => Adapter restarts
Test: SetLLPrivacy on TurningOn => Adapter restarts and eventually on
Test: SetLLPrivacy on TurningOff | Restarting | Off => No-op

Change-Id: Id9a1c06d1e71bca932ee59f1c3810c85404ca794