Revert "[le audio] Add counter metrics for le audio health status"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/911dfd18-5108-4ab9-b1d4-fe3dd9d626a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb8d7b849c192e13872c058cfcf9896425eff361