Revert "[le audio] Add allow list check when le audio device con..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/911dfd18-5108-4ab9-b1d4-fe3dd9d626a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5356d564f937d6a85e6ba6869bfe246f1752d15b