Revert "Health based action: Enable first native flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/911dfd18-5108-4ab9-b1d4-fe3dd9d626a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33d3132463280226eb9a1f8e0d06041ea67553b2