floss: hcidoc: modify profile storing mechanism
The same ACL connection might have several parallel connections to
the profile. We need to differentiate between the connections.
Therefore, this CL introduces a ProfileId, which is a custom field
for this purpose. For profiles based on L2CAP, CID can be used as the
ProfileId, and for RFCOMM based profiles, one can use the dlci.

Bug: 306108678
Test: m -j
Test: Manually, check HFP detection doesn't regress.

Change-Id: I2bf40d65170f8ebfcbcbf44465eb0a446f79d086