Revert "Remove test with ASAN enabled from hwasan-presubmit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a25dd1ef-304b-48ab-9c73-12b5e5150984). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaef1b140b88e1b35ee3b56cb03a03599431beaf8