Revert "Move all device-tests to general-tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a25dd1ef-304b-48ab-9c73-12b5e5150984). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia591fa03ff7019b26e6187378bda8a23ebb4c005