Revert "Use one definition for Octet16"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d293d22b-8668-4b96-8fba-fd1191b3358c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72fb5553c8bff631f7b26f09b7a48b0c2454a490