Refactoring: Audio Routing Handler (Step 5)
This CL refactors connection and activation events to be handled in
the AudioRoutingHandler. This will remove the mLock and migrate the
variables guarded by the mLock into the handler.

It will also remove the priority of hearing aid devices so that they
are activated at the same level as other BT devices.

This CL refactors the fallback device selection logic.
By removing the priority of hearing aid, it becomes much simpler.

Bug: 299023147
Test: atest BluetoothInstrumentationTests
Change-Id: I95bcaeba37e88bf88f3136858914ee0385c05dcb