Revert "[le audio] Add allow list check when le audio device con..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50b4da63-e317-431b-bedc-f84a97451f27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I724784a04e815cd4cde8ed33ef76df0640bdea76