Revert "[le audio] Add counter metrics for le audio health status"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50b4da63-e317-431b-bedc-f84a97451f27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9463c41d58a1fb1f816c9112ff3b7f42d24af2ea