Add startScan() with PendingIntent tests
Create PendingIntentScanReceiver class that is statically registered in
the manifest to receive scan result broadcasts. This class can also
be used as a dynamically registered broadcast receiver.

Create startBleScanWithPendingIntentAndStaticReceiverAndCallbackTypeAllMatches
test that scans using the statically registered
PendingIntentScanReceiver.

Create startBleScanWithPendingIntentAndDynamicReceiverDoesNotReceiveScanResults
test that scans with a dynamically registered PendingIntentScanReceiver.
Note that this functionality appears to be broken currently.

Bug: 306062963
Test: atest BumbleBluetoothTests:android.bluetooth.LeScanningTest
Change-Id: Ia34f8e08002bc66680ca95f7fb6428a54617a182