Revert "BluetoothProfileConnector: Remove mProfileName"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7a13fde-5d18-4017-ac82-aa288ef7aacd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fcbfdcc625cf2163dd85579a3f293ab7ca82c23