Revert "Make BluetoothProfileConnector final"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7a13fde-5d18-4017-ac82-aa288ef7aacd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd666e55ad519efa312b58b07e939af5bf87e236