Floss: Fix not able to answer 2nd incoming call
The "Dialpad" application opens the SCO link just after sending the
incoming call UHID output report, which triggers the sending of a +CIEV
event as if the call was answered in "Dialpad".

To prevent this unexpected interaction with the telephony the +CIEV event
should be sent only when there is no telephony call in progress.

Bug: 310065944
Test: Conduct the following manual tests.
  pair Bluetooth Headset
  login to "Dialpad" via https://dialpad.com/login
  add Headset to "Dialpad" call control by following step:
    Profile -> Preferences -> Hardware Settings -> Call controls
    -> Add Headset -> Select headset and click connect.
  call to "Dialpad" with phone
  check there is a incoming call
  answer call on headset
  end call on headset
  perform a second call to "Dialpad" with phone
  check there is a incoming call
  answer call on headset
  end call on headset

Change-Id: I71f98a2f15bf61e9f9a1741aed7d96540edf2858