Revert "Make BluetoothProfileConnector final"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1f0df12-415d-4c6d-b7f8-4bfd3b9638ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibddc569aac40211fad950a95a53de2fa9ba3cbc6