Fix the order of key notif after conn enrypt.
After the BLE connection is encrypted, HOGP device should be able
to send key notification if it is enabled in CCCD. However after
reboot, the UHID device is not opened to be able to handle the
notification. So we need to queue the UHID_INPUT events when UHID
device is not opened and send them all at once when UHID is opened.

See https://kernel.org/doc/Documentation/hid/uhid.txt

Bug: 299192582
Test: Key sent after encryption can be handled.
Test: atest net_test_btif_hh

Change-Id: Ib1223a91bf8275035df8851831e15815bd8d34ff