Delegate bta_hh_co_data to the thread uhid is opened.
According to spec, HOGP device can send key reports to the client after
the connection is encrypted if its corresponding CCCD is enabled. Android's
HOGP implementation hasn't opened the bta_hh_device_t in the BT interrupt
thread if the notification is just after the connection.

To resolve it, invoke the bta_hh_co_data in the same thread as uhid is
opened so these operations are serialized.

Bug: 299192582
Test: Can reconnect with HOGP device
Test: atest net_test_btif_hh
Test: banchan com.google.android.btservices mainline_modules_arm && UNBUNDLED_BUILD_SDKS_FROM_SOURCE=true m apps_only dist

Change-Id: I4195c21146a9757fa9b9e3d8503736e857a0b6bc