Revert "Move BluetoothProfileConnector out of profile proxies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce97cd35-096f-437a-84b5-855e3ba440ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I339534638d38470cb5778a73f6e1e49add92a8a3