Revert "Counter metric for calls to closeProfileProxy on differe..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce97cd35-096f-437a-84b5-855e3ba440ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48919978ead1698e32e26da985ec26222b036f17