Clean up correct state machine
MceStateMachine#doQuit is called inside tearDown(),
which call MceStateMachine#onQuitting,
which cleans up the device.

During the testing, a new state machine can be assigned
to the same device ID, which could be cleaned up by
the previous state machine.

By accepting the state machine instance, this change
prevent this.

Bug: 308622269
Test: atest MediaClientServiceTest
Change-Id: I2310845445a2d040a81331147f42b11cbe5b92a5