Revert "Counter metric for calls to closeProfileProxy on differe..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8a1db77-1950-4100-9655-24d85668c996). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1bc3efee4482c6f57a91effd08e84a12bf61a782