Revert "Counter metric for calls to closeProfileProxy on differe..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6292ad6a-0558-4a02-a684-8755603c0bae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c3b300b05153e8f68e8df57e2c2d8861ff3b156