Revert "Move BluetoothProfileConnector out of profile proxies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6292ad6a-0558-4a02-a684-8755603c0bae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a90b421d97aa859e75ab5064e33e6f8d96b4254