Revert "Properly include stack::sdp_status::tSDP_RESULT"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72dd9dd1-569e-4e4e-b4ba-cfb2682259f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I191c7d034c6954b2ec9c68ff6b9b8ba248594b9e