Revert "Properly include stack::sdp_status::tSDP_RESULT"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54468683-91cc-46b6-89eb-5c3a4a34ae4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I932f9a549163f5661f635d42fd5e56e822cb2949