Revert "Remove duplicated test bt_host_test_bta"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d3a744a6-4da7-4026-89eb-ac6e13cc7120). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a4159a0fc9164f21c208e5bec054b2ebac73310