hearing_aid: Do a direct connection if another ear is connected
On connection, if another device from the set is connecting,
transform the background connection to a direct connection to allow
to scan more aggressively and connection as fast as possible.

On a peer disconnection, if another device from the set is still
connected issue a direct connection. If this direct connection fails
it will be transformed back into a background connection.

Bug: 308032317
Test: TODO
Change-Id: Ie4fd9885fbc57dd608a94f173c21d0a1130efa09