hearing_aid: Do a direct connection if another ear is connected
On connection, if another device from the set is connecting,
transform the background connection to a direct connection to allow
to scan more aggressively and connect as fast as possible.

On a peer disconnection, if another device from the set is still
connected issue a direct connection. If this direct connection fails
(after 30s) it will be transformed back into a background connection.

Bug: 308032317
Test: manual, open Wireshark and look at the LE Extended Create Connection
  Scan Interval + the time it takes on the UI when:
  - taking the HA out of the case
  - putting one HA back in the case and taking it out
  - putting one HA back in the case and wait 30 seconds
Change-Id: Ie4fd9885fbc57dd608a94f173c21d0a1130efa09