le_impl: Fix direct connection for device already in acceptlist
create_le_connection relies on a side effect of
add_device_to_connect_list to disarm and re-arm itself.

If a device was already in the acceptlist, add_device_to_connect_list
will do nothing and will not re-arm the connection.

This commit adds some logic to explicity handle this case, by
disarming the le_impl manually and marking a flag to re-arm
it afterwards.

Bug: 308032317
Test: atest --host LeImplTest#direct_connection_after_background_connection
Change-Id: I9c7121457bc49f1fbe4c64f588f4cce22ec4e510