Revert "leaudio: Fix for race in late connect scenario"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91654550-7e2a-4272-93f8-043c01f9d579). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I56e153d07c4eeb872b4acc307521d1c4a49fc6fb