Workaround for the audio routing issue from LEA to classic
This is the revert patch of aosp/2733933 to resolve the audio routing
issue. As the app use BluetoothAdapter.setActiveDevice to switch the
routing, the flow doens't be integrated with ActiveDeviceManger. The CIS
disconnected triggered with delay that cause the A2dp start before CIS
disconnected and ISO data removed. Most of the controller couldn't
handle this overlap.

Bug: 307481636
Bug: 284432213
Test: atest A2dpServiceTest
Change-Id: Ie1ebfb5bb939fcc26536af67708b2457ff84245e