Revert "bta: Update legacy bta::discovery tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ecde1e1-3a1b-4a8c-82af-61fe71053763). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I500d5c49c900c832a14a54e2af9d71111eb7808a