LeAudio: Clean-up the queued broadcast data
Due to static life time, the state of queued broadcasts is not
cleaned-up. There is no need for static life time for it as the
instance is not passed to any other module and is accessed from
static function over the static service implementation instance.
We can limit its life time to the instance life time.

This also fixes uninitialised member variables.

Bug: 295972694
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Change-Id: Iaa19a10af51edaf89108affda19a7582c3f78d95