LeAudio/Broadcast: Simplify the broadcast queueing mechanism
QueuedBroadcast class is redundant, as there is already a container
for the entire broadcast configuration (BroadcastStateMachineConfig).
We can simply queue this config and pass it later to the state machine
construction when needed. Additionally, the ISO state monitoring logic
did not seem to fit the QueuedBroadcast responsibilities and as an
ortogonal functionality can use a single boolean variable to
store the current state. This gives us room to greatly simplify
and reduce the amount of boilerplate code.

Bug: None
Test: atest --host bluetooth_test_broadcaster
Change-Id: I115120d33e33302be5127218b02b217cfccd1a32