Revert "net_test_btif_rc: Source src/btif_util.cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ac5a9398-0039-4028-8dd3-467ccdc4c8eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I287a49bf329aafa430dd1903cc4342535dda1564