Revert "GattClientTest: User wrapper in all test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0c3f422-e9f6-4d73-bc02-4cccad74c2b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0895aa416049694fa186850cc12b4985b3f6655