Floss: add workaround entry to insert call when sco start
Bluetooth headsets that require CIEV commands to provide sound when HFP SCO starts can cause problems when developing Bluetooth telephony features.

This is because the workaround requires inserting one call into the call list, which can make it difficult to maintain the correct call state.

This can lead to bugs such as users being unable to answer calls the second time they are called, or being unable to answer calls while other audio streams are playing.

To address this issue, we have added an entry in interop that allows us to only apply the workaround to headsets that require CIEV commands based on their address.

Bug: 310065944
Test: atest bluetooth_test_gd
Test: disable bluetooth telephony, make sure the call exist when sco
start
Test: enable mps qualification mode, make sure the no call exist when sco
start
Test: Conduct the following manual tests.
  enable bluetooth telephony
  add Jabra Evolve2 address into /var/lib/bluetooth/interop_database.conf
  check the log show
  2023-12-06T07:02:03.719513Z WARNING btadapterd[19584]: bt_device_interop:  device/src/interop.cc:1355 - interop_database_match_addr: Device xx:xx:xx:xx:5d:59 is a match for interop workaround INTEROP_INSERT_CALL_WHEN_SCO_START.

Test: Conduct the following manual tests.
  pair Bluetooth Headset
  login to "Dialpad" via https://dialpad.com/login
  add Headset to "Dialpad" call control by following step:
    Profile -> Preferences -> Hardware Settings -> Call controls
    -> Add Headset -> Select headset and click connect.
  call to "Dialpad" with phone
  check there is a incoming call
  answer call on headset
  end call on headset
  perform a second call to "Dialpad" with phone
  check there is a incoming call
  answer call on headset
  end call on headset

Change-Id: Iea1029f618dcd1d204902cd72fcf234473b74144