Revert "Backport "Reorganize the code for ..." to AOSP"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47c86acb-3562-4507-8134-0bd7a384db0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id27dec14aef27ed33951a45086069548dc237dea