Revert "Backport "Reorganize the code for ..." to AOSP"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a20ada1-3e86-4438-ae5b-8c92c895d61d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifad8ff66ff9a9b8ba12f40d6b9976ce09a749344