Revert "Backport "Reorganize the code for ..." to AOSP"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c06b2bbb-0c5d-4ce3-b92a-b3ae00951e40). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibba788fb5687bcb8795480a8c03c5da5434bda52