Fix premature GATT DB lookup
As part of a change made to support Robust Caching, our code makes an
early query against the peer's GATT database hash characteristic,
skipping the check only when the peer is known not to support Robust
Caching. However, the check is still made even on initial connection,
when the peer's support for Robust Caching is unknown and no database
entry exists for it.

Skip the check if the peer's support is unknown.

Bug: 295365121
Test: atest net_test_stack_gatt
Change-Id: I0d3e406b0ce70edd11cdab0bb355cc448e40fed7