Floss: Handle Unknown Connection Error as if Disconnnect Complete
When conntroller replies |Unknown Connection| (0x02) error in command
status event, although it could be a controller issue, we can still
handle it as |Disconnect Complete| HCI event so that the stack could
properly cleanup the connection state.

Currently we only add this policy to |LE Read Remote Features| because
we only observe the issue on this command.

Bug: 292116133
Bug: 316928664
Test: Tested with byra/Taniks and MX Master 2S, which is a known
combination to trigger |unknown connection| described above.
Verified that after the fix, the device can be reconncted.
Test: mma -j32

Change-Id: I2f24c2d35fd6509ba1d1c3c3b8d53410d297aa97