Allow constructing BluetoothDevice with anonymous address type
Define private value for anonymous address type defined in Bluetooth
Core 5.4 spec.

Value added to IntDef is private in BluetoothDevice class because it should
not be allowed to be used to create BluetoothDevice object by
application, and is passed directly from native in onScanResult
callback. Package private constructor not an option as is constructed from different package in the regression being addressed by this CL.

Test: manual: scan anonymous advertisement and ensure scan result is
delivered to application
Bug: 296988500

Change-Id: I8168c370ec63f48acd535ea4ef05c6d5e58c2a2d