Revert "Add new flag for scan timeout reset"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83c407d3-5e72-4c3a-b8fc-ae68076f13f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2395d3555370712bba03214052e959990995d2be