Don't return from inside log tag enforced blocks
A few places in our code have return statements guarded by whether or
not DEBUG or VERBOSE logging is enabled. This is very dangerous and
allows for logic to be determined by log level settings. These are
certainly not intended and the return statements should be moved _out_
of the blocks, and be based on the conditions that caused that log.

Tag: #stability
Flag: EXEMPT, logical no-op, as logging flags were defaulted to true
Bug: 315046089
Test: atest HeadsetClientStateMachineTest.java
Test: atest LeAudioServiceTest.java
Change-Id: I2ae4e0a4c8042c7792460358ae98af367a65195a