Revert "bt_target:[26/37] IWYU btif:: device::"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a87d8aa0-208d-4cb8-9227-22f4c389dd22). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ea0d07658c20ac0355ebd70c44137aa8d1714fd