Revert "Remove internal_include directory path from build files"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a87d8aa0-208d-4cb8-9227-22f4c389dd22). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89cd30464d2e5f497e93f6a3287fe10f7d4cab16