LeAudio: Support vendor configurations in control point operations
With this change we store the raw data in ctp_codec_conf instead of
storing the source LTV formatted codec configuration, from which we
generated the raw data at the later stage when the control point is
written.

By moving the raw data generation to an earlier stage we prapare the
ctp_codec_conf structure to pass the opaque vendor codec specific data,
and not just the LTV data format used by the LC3 codec.

Bug: 308427707
Test: atest bluetooth_le_audio_test
Flag: EXEMPT; refactor for the multicodec support, only moved the data buffer generation to an earlier stage, and the end result is the same - confirmed with the unit tests.
Change-Id: Ieec55300951b94c8f1f2d2ee476fe9606d18cb58