floss: Generalize mSBC decode buffer
When we begin transmitting data to the controller, the
controller generates some dummy data to prevent overrun.
Occasionally, the size of this dummy data is not aligned with
the mSBC audio packet, which disrupts the current assumption that
read/write pointer offsets are always multiples of 60 or 72.
This misalignment impedes our ability to enqueue the decode buffer.

The patch addresses this issue by introducing a more generalized
assumption for the mSBC decode buffer, enabling it to handle
such cases effectively.

Bug: 314266464
Tag: #floss
Test: m Bluetooth && Ran HFP tests and verified HFP WBS
works on devices with SCO packet size 60 and 72

Change-Id: I89b97617845f6685fb1332c619f3b64c2cc074c7