McpService: Align with AVRCP behavior.
With this patch, base media control opcodes are always send up to the
player and are always exposed over SUPPORTED OPCODES characteristic.
This is to match AVRCP and workaround the issue with player being to
registered on time when LE HS connects when player is already started.

Bug: 312324138
Test: atest MediaControlGattServiceTest
Tag: feature
Flag: Exempt, LeAudio feature to match BR/EDR functionality
Change-Id: I1520712beca2abff693cce1b1b380e021c8c6cae