Specify aconfig_declaration modules in java_sdk_library
java_sdk_library modules should specify the aconfig_declarations modules
in order to generate the exportable stubs (i.e. the stubs that exclude
the disabled aconfig flags) by providing the information about the
aconfig flag values.

The exportable stubs will be used when generating the module sdk
snapshots and to generate the hiddenapi flags, in order to prevent
exposing the disabled apis in runtime.

Test: m framework-bluetooth
Bug: 317290555
Change-Id: I86d6142dfa1cd7c1a73718b615f2f89a36b2416b