LeAudio: Fix redundant audio hal stop
Trivial change - moved calling StopSession() on self from Cleanup()
method, to the parent object before it calls this Cleanup().
The parent tracks the interface state, so it calls StopSession() only
when it is actually running.

This fixes the additional StopSession from Cleanup() on the non-started interface:
 luetooth: packages/modules/Bluetooth/system/bta/le_audio/audio_hal_client/audio_source_hal_client.cc:362 - Stop:
 bt_stack: [INFO:le_audio_software.cc(313)] StopSession sink
 BTAudioClientLeAudioStub: packages/modules/Bluetooth/system/audio_hal_interface/le_audio_software.cc:103 - Cleanup: HAL transport: 0x01, is broadcast: 0
 t_stack: [INFO:le_audio_software.cc(313)] StopSession sink

Bug: 320275987
Test: m com.android.btservices; tested manually
Flag: EXEMPT; trivial change
Change-Id: I1e29d42fb1933470625964772b24a1d404a7aca6