Fix bta_dm_disc stuck in cancelling state
When service discovery is cancelled, we still expect to receive some sort of
result either from SDP or GATT, and won't move to any different state
until we do.
If service discovery fails to establish GATT connection, or if it was to
fail i.e. due to peer disconnection, we shoudl move on to next device
trying to discover services.

Bug: 318580354
Test: Try bonding, disconnect device while GATT service discovery is
scheduled but not yet done

Change-Id: I7c59701831d0b9288dea9eea375769db357fb7bc