[le audio] Switch active device group during broadcast
If broadcast is ongoing, the update active device request should update
the cached unicast fallback device group id, so that when broadcast
stopped the device can switch to the correct group.
And in this case, the device don't need to set the active device during
broadcast.

This change is checking if there is active broadcast which is by nature
behind the feature flag.

Bug: 319939252
Bug: 316005152
Test: atest LeAudioServiceTest
Test: manual test with new UI and switch active device during broadcast
Change-Id: If64c5065f0fdd03eb8b33811fc66e06099ba1fcc