Revert "Baseline NewApi issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b9371dba-896d-4cfd-a432-867b3153a39f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I949c3af8e6fa76cd4a7f71e499a33193c77fe12e