LeAudio: Configuration provider dependency inversion - Part 2
Part 2: unified SWB API

Codec Manager will provide a common interface to managing codecs, codec
configurations and their porperties (like SWB support), regardles of the
audio session type and codec location. This change moves the SWB checks
under the Codec Manager API so that the Unicast code would simply call
to the manager and not have to check for codec location and take
different actions based on that.

Bug: 308428860
Bug: 295972694
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; refactor for the multicodec support - verified with unit tests and manual testing
Change-Id: I54da33bc57d72088c9b04e4e41a784d7a31d6fb7