btif: fix BTA_DM_AUTH_SMP_CONN_TOUT handling
In BTA_DM_AUTH_SMP_CONN_TOUT, we want to forget device, if bonding or
initial service discovery is in progress.
Otherwise, we want to just report issue with encryption.

Up till now, the logic for detecting if device is in process of
establishing bond was not correct. This resulted in BondStateMachine in
java being stuck until phone restart.

To fix this, use BTIF pairing state, and service discovery information
instead of old method which considered devices immediately after key
exchange as bonded.

Test: manual bonding test, have remote device disconnect less than 500ms after bonding on LE transport, and verify bonding can be restarted without restarting the phone

Bug: 309141668
Flag: exempt, trivial fix of corner case, tested manually
Change-Id: I68746e8172b15b35b68ec5e65e54d6277b022076