[le audio] Remove source should trigger update first if BIS is synced
If the sinks are actively synced to source, the remove source request will trigger modify source operation if PA synced.

We should also do the same if PA is un-synced but BIS is synced since some buds might choose to unsync PA and only sync to BIS to save power.

Bug: 321292454
Bug: 316005152
Test: atest BassClientServiceTest BassClientStateMachineTest
Change-Id: I357de25a8a12cc6489d7475c3b51fbaa3f9ead57