LeAudio: Configuration provider dependency inversion - Part 1
Part1: Move broadcast configurations to Codec Manager

Codec manager should be the only entry point for the BAP code, which in
turn should not care for the codec location and from where the
configurations are provided.

This is part of the refactors preparing the code base for the multi-codec
support.

Bug: 295972694
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; refactor for the multicodec support - verified with unit tests and manual testing
Change-Id: Ibfe28cba5419d92af41ede4117a1f7d72dbbc8c4