Fix BluetoothGatt busy state after failure
Previously, when BluetoothGatt#writeCharacteristics
failed early, the `mDeviceBusy` incorrectly remain
`true`. This prevented subsequent operations
on the BluetoothGatt instance.

This change ensures the `mDeviceBusy` is properly set to
`false` after a failure is returned by the service,
allowing for continued use of the instance.

It also adds a test for ensuring the behavior.

Bug: 322580271
Bug: 322509603
Test: atest GattClientTest

Change-Id: I6aea853745e15aca5c4caf75037e680592e74c81