csis: Use opportunistic connect when bonded
csis: Use opportunistic connect when bonded

When device is bonded, use an opportunistic connect for CSIP, same as
for VC.
We need CSIP to keep connection alive only during bonding - and this
still is a case.

With this approach it is OK to remove to "autoconnect" flag logic.

Bug: 314229105
Flag: Exempt, regression covered with unit tests
Test: atest bluetooth_csis_test
Tag: #feature
Change-Id: I7a07ff699300ad5a8a34408d87cf72069a4432d3