[le audio] Remove the fallback device settings from service start/stop
This CL reverted the change in aosp/2942290 and log exception if the
setting operation failed.

In case of switching multi-user, there might be a race condition window
that activity manager user setting is not updated yet but leAudioService is
restarted with the previous user. Then this will fail with
SecurityException.
We need to avoid calling this from service start/stop, and let LEA
broadcast operations handle this setting properly.

Bug: 319939252
Bug: 316005152
Test: atest LeAudioServiceTest
Test: manual test with new UI and switch active device during broadcast
Change-Id: Ie3db50592751af58b6881e47d7f0e80d3d2198d6