Move librootcanal_rs from whole_static to static
Declaring a Rust staticlib as a whole_static_lib dependency can lead to
duplicate symbols in the final linkage of dependents if they contain
multiple whole_static_lib Rust libraries in their dependency tree.

This moves librootcanal_rs to be a normal static_lib dependency, and
requires it to be redeclared as a dependency in libbt-rootcanal
dependents.

Bug: 254469782
Test: m netsimd
Change-Id: Ied4ab866c8c54e521aa717b2025b841dc9894445