Revert "acl_api apply sysprop for random address rotation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43e374d9-78eb-4ed6-9a78-cb460b041500). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I823f10f73547bacb70eaf73d3a1bc4f94f855843