Route A2DP Traffic
Problem: When both the a2dp profiles are enabled concurrently, the
method calls to btif_av is unaware which profile to redirect the
incoming/outgoing traffic.

Solution: Pass the type of profile that the client is interested
in. BTIF AV redirects the traffic based on the type of the profile.

Test: atest net_*

Bug: 315334681
Bug: 321806163
Change-Id: I374773e3274708dc00242bf260bc2a5fbca549b4