Revert "Security: Register for the pairing events"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ce49bec-b3bb-44b6-bf51-950b33b35764). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida4d6580b0d1a0b634d2015e7e67082d38906a2a