LeAudio: ConfigureAses improvement
We should be using the current number of connected device as a device
counter in the remote device ASE configuration procedure instead of the
number of devices from the cached configuration. Even though these two
match currently, in future the configurations may not have the device
count requirement at all.

Bug: 308428860
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Flag: EXEMPT; Part of the multi-codec refactor, covered by the unit tests
Change-Id: I955714eb57803d963cd39c17779be11d78fa3d5a