AV Suspend from source device should not suspend a2dp source stream
Problem: While bluetooth's source profile is streaming audio to a
remote sink device and in parallel a source device connects to the
sink profile. The source device may send AV Suspend event
(for sink profile) to suspend the audio stream. Currently, it
results in the source profile to suspend the audio stream.
Also, when the a2dp stream goes idle, there is a possibility there
is no active device resulting in the idle event to be dropped.

Solution: Add additional check for which profile the stream is being
suspended/idle event for.

Bug: 325528613
Bug: 321806163
Test: atest net_*
Change-Id: I6d05dd28dc3bef068d95d3aa4b13aa4326d351a0