floss: start resume if suspend is aborted
If the suspend is aborted in the powerd, the SuspendImminent signal is
cancelled after receiving SuspendDone. Bluetooth will start the resume
flow after the suspend is ready.

Put LE Rand to the end of the suspend/resume flow, so the callback can
clean the states properly.

Tag: #floss
Bug: 324424415
Test: shift+search+L
Test: mma -j32
Flag: EXEMPT, floss change for ChromeOS only.

Change-Id: Ibdb087d2d2e5aa9938c2debd8f74f4208b6f41b6