Floss: Hcidoc: Add external controller signal
Some users plug an external controller, if the stack somehow use that,
it is likely to fail in various situation as it wasn't tested.

This adds a signal to detect if the chosen controller is an external
controller by checking if the local name and manufacture are in the
known list.

Bug: 325403872
Test: m -j
Test: Using report 91563648247, verify every read local name is reported
Tag: #floss
Flag: EXEMPT, Floss only change

Change-Id: Ia25c4f8deaaa20d4529d412938c3cc1fdcf915c0