Fix parsing UUID on EIR
It is assumed that btm_eir_get_uuid_list returns the number of UUIDs
in EIR. However, the implementation is incorrect, and currently it can
only return 0 or 1.

There is no straightforward way to supply the correct amount of UUID
and the location of the UUIDs without introducing additional memory or
computational cycles. Therefore, this CL modify the caller of that
function to update the expectation, rather than updating the function
itself.

Bug: 291973444
Test: m -j
Test: Verify EIR UUID is correctly logged
Change-Id: I8bba7d9e5a947a8dfee0375b6913d410e6d791f7