[WIP, BCST needed] OppService: Fix IndexOutOfBoundException in UpdateThread
Due to the timing change between ProfileService#stop() and
ProfileService#cleanup(), the main thread and the UpdateThread were
accesing the same list (mShare) concurrently.

This CL calls the stopInternal() in BluetoothOppService#stop() so that
the list is cleaned up after the UpdateThread is stopped.

TODO: Flag change
TODO: Run test after enabling the test

Bug: 321897599
Bug: 326135849
Test: atest BluetoothInstrumentationTests
Change-Id: I35d6507e32125aac63b38d0f26b8d8f1c62a1ac6