Revert "Using tAclLinkSpec in place of RawAddress in HID host"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0262a4e5-cca1-4003-8606-2476f8965733). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15b8abd2ce1cad57ed8417baccca188d1b74e0d1