Let the Log framework enforce log levels everywhere
It's a misconception that the Log framework does not enforce log levels
against the log lines written, and that all calls to Log.<level>() must
be protected with Log.isLoggable() calls (or something similar) to
govern what's written to the log buffer and prevent log spam.

In all reality, the log framework _does_ actually enforce log levels,
but only if any of the following are set (applied in this order):
 - A explicit tag (adb setprop log.tag.<TAG> <level>)
 - A process default (made with a native call)
 - An Android wide default (adb shell setprop log.tag <level>, or build
   time set as I understand it but nobody does)

If none of these are set, a default minimum of VERBOSE is used, which
means "log everything." Note that this explicitly disagrees with the
return values of Log.isLoggable(), which uses a default minimum of INFO
in the case none of those three values are set, which greatly adds to
the confusion here.

The problem is that nearly no one satisfies any of these criteria, and
its perpetuated the idea that we all need to do extra work.

As further proof of the claim, the Java Log framework leverages native
calls for all log line writes. Tracing the code, when call Log.d(),
you get the following chain of calls-- println_native() -> (JNI) ->
android_util_Log_println_native() -> __android_log_buf_write() ->
__android_log_is_loggable(prio, tag, ANDROID_LOG_VERBOSE). It clearly
calls __android_log_is_loggable(), which has an implementation that
checks the above values, in the order mentioned!

So, with the addition of a process minimum log level in an earlier
change, we now know our process default will be enforced against all
log writes made, regardless of Java calls to Log.isLoggable(). This
means our log lines no longer need to be protected, and we can let the
framework properly do it's job.

This change removes the following patterns:
- Static variables that represent cached log levels, both with hardcoded
  "true" or "false" values, and Log.isLoggable(TAG, <level>)
- Protected log lines, either against Log.isLoggable() directly, or the
  above static variables

This brings several important benefits:
- Runtime log level enforcement for all builds: Just set a tag on a
  build and logging will instantly begin without the need for an adapter
  power cycle. You also don't need to make a build change to allow
  debugging since that pattern is removed as well.
- Simple default log level control: Adjusting the process minimum
  default will impact the log level for _all_ tags in the stack, without
  having to explicitly know/set the tag. We've done this with a special
  tag, "log.tag.bluetooth"
- Lower complexity/Less code: Just call to make a log. No need for extra
  code or abstractions.

Tag: #refactor
Flag: EXEMPT, logging only change and a logical no-op
Bug: 315046089
Test: m com.android.btservices
Change-Id: I25d5b8e28dc29970c77bff65525789aabe191fe6